backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
alpaca-trade-api
yfinance
httpx
sqlalchemy
aiosqlite
pydantic[email]
bcrypt
python-jose[cryptography]
mistralai
